[DEV-2.0] Refactor DB codebase to act as an isolated microservice
The DB now runs as a seperate service on the new port: 9000. The following folders/files have been created:

- requirements.txt: contains requirements to install to virtual environment, prior to starting the service
- database.py: where the DB service runs, this is the file you run using Python to start the service
- db_connection.py: contains DB connection variables, used merely as an import wherever necessary
- repositories/chat_repo.py: contains DB logic for reading and creating chats
- repositories/user_repo.py: contains DB logic for reading and creating users
- repositories/session_repo.py: contains DB logic for reading and creating sessions

Note that the new folder named `repositories` comes from using the repository design pattern in this codebase. The underlying files
are basically routed through an API, to adhere to our microservice-based architecture.
