+++ README.md
@@ -1,2 +1,7 @@
 # speakwrite-db
 Database repository
+
+To start up, create a Python virtual environment and then:
+```bash
+pip install -r requirements.txt
+python3 database.py
